+++ flowz/channels/__init__.py
@@ -1,5 +1,21 @@
 from __future__ import absolute_import
 
-from .core import (ChannelDone, Channel, ReadChannel, MapChannel, FlatMapChannel,
-                   FilterChannel, FutureChannel, ReadyFutureChannel, TeeChannel,
-                   ProducerChannel, IterChannel, ZipChannel, CoGroupChannel)
+from .core import (
+        Channel,
+        ChainChannel,
+        ChannelDone,
+        CoGroupChannel,
+        FilterChannel,
+        FlatMapChannel,
+        FutureChannel,
+        GroupChannel,
+        IterChannel,
+        MapChannel,
+        ObserveChannel,
+        ProducerChannel,
+        ReadChannel,
+        ReadyFutureChannel,
+        TeeChannel,
+        WindowChannel,
+        ZipChannel)
+